Module oak_containers_launcher::proto::oak::attestation::v1

Modules

endorsements

Nested message and enum types in Endorsements.

Structs

ApplicationKeys

Keys used by the application to derive encryption session keys and to sign arbitrary data. Each of the certificates contains the final layer’s measurement as additional claims.

ApplicationLayerEndorsements

CbEndorsements

CertificateAuthority

Message for passing embedded certificate authority information between layers. Will never appear in the evidence that is sent to the client.

ContainerLayerEndorsements

DiceData

Message that is sent between DICE layers.

Endorsements

Endorsement contains statements that some entity (e.g. a hardware provider) vouches for the integrity of claims about the TEE or the software running on it.

Evidence

Attestation Evidence used by the client to the identity of firmware and software running inside a Trusted Execution Environment.

KernelLayerEndorsements

LayerEvidence

DICE layer evidence containing a certificate signed by the previous layer.

OakContainersEndorsements

OakRestrictedKernelEndorsements

RootLayerEndorsements

RootLayerEvidence

Evidence generated by the Layer0.

SystemLayerEndorsements

TransparentReleaseEndorsement

Endorsement for binaries from the Transparent Release process. <https://github.com/project-oak/oak/blob/main/docs/release.md>

Enums

TeePlatform

This proto defines the layered DICE Attestation Evidence.