Module oak_containers_launcher::proto::openmetrics

Modules

counter_value

Nested message and enum types in CounterValue.

gauge_value

Nested message and enum types in GaugeValue.

histogram_value

Nested message and enum types in HistogramValue.

metric_point

Nested message and enum types in MetricPoint.

state_set_value

Nested message and enum types in StateSetValue.

summary_value

Nested message and enum types in SummaryValue.

unknown_value

Nested message and enum types in UnknownValue.

Structs

CounterValue

Value for COUNTER MetricPoint.

Exemplar

GaugeValue

Value for GAUGE MetricPoint.

HistogramValue

Value for HISTOGRAM or GAUGE_HISTOGRAM MetricPoint.

InfoValue

Value for INFO MetricPoint.

Label

A name-value pair. These are used in multiple places: identifying timeseries, value of INFO metrics, and exemplars in Histograms.

Metric

A single metric with a unique set of labels within a metric family.

MetricFamily

One or more Metrics for a single MetricFamily, where each Metric has one or more MetricPoints.

MetricPoint

A MetricPoint in a Metric.

MetricSet

The top-level container type that is encoded and sent over the wire.

StateSetValue

Value for STATE_SET MetricPoint.

SummaryValue

Value for SUMMARY MetricPoint.

UnknownValue

Value for UNKNOWN MetricPoint.

Enums

MetricType

The type of a Metric.