Enum oak_containers_launcher::proto::openmetrics::MetricType

#[repr(i32)]
pub enum MetricType {
    Unknown,
    Gauge,
    Counter,
    StateSet,
    Info,
    Histogram,
    GaugeHistogram,
    Summary,
}

The type of a Metric.

Variants

Unknown

Unknown must use unknown MetricPoint values.

Gauge

Gauge must use gauge MetricPoint values.

Counter

Counter must use counter MetricPoint values.

StateSet

State set must use state set MetricPoint values.

Info

Info must use info MetricPoint values.

Histogram

Histogram must use histogram value MetricPoint values.

GaugeHistogram

Gauge histogram must use histogram value MetricPoint values.

Summary

Summary quantiles must use summary value MetricPoint values.

Implementations

impl MetricType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of MetricType.

pub fn from_i32(value: i32) -> Option<MetricType>

Converts an i32 to a MetricType, or None if value is not a valid variant.

impl MetricType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.